# Updates to sync_float_amax_history